autenticacion por tipos de usuario

+++ laravel/routes/api.php
@@ -20,18 +20,18 @@ Route::post('/logout', [AuthController::class, 'logout'])->middleware('auth:sanc
 Route::get('/user', [AuthController::class, 'user'])->middleware('auth:sanctum');
 
 /* ------ AREAS ------ */
-Route::get('/areas', [areaController::class, 'index'])->middleware(['auth:sanctum', 'abilities:server-admin']);
+Route::get('/areas', [areaController::class, 'index']);
 Route::get('/areas/{id}', [areaController::class, 'show']);
-Route::delete('/areas/{id}', [areaController::class, 'destroy']);
-Route::post('/areas', [areaController::class, 'store']);
-Route::put('/areas/{id}', [areaController::class, 'update']);
-Route::patch('/areas/{id}', [areaController::class, 'updatePartial']);
+Route::delete('/areas/{id}', [areaController::class, 'destroy'])->middleware(['auth:sanctum', 'abilities:server-admin']);
+Route::post('/areas', [areaController::class, 'store'])->middleware(['auth:sanctum', 'abilities:server-admin']);
+Route::put('/areas/{id}', [areaController::class, 'update'])->middleware(['auth:sanctum', 'abilities:server-admin']);
+Route::patch('/areas/{id}', [areaController::class, 'updatePartial'])->middleware(['auth:sanctum', 'abilities:server-admin']);
 
 
 /* ------ DOCTORES ------ */
 Route::get('/doctores', [doctorController::class, 'index']);
 Route::get('/doctores/{id}', [doctorController::class, 'show']);
-Route::delete('/doctores/{id}', [doctorController::class, 'destroy']);
-Route::post('/doctores', [doctorController::class, 'store']);
-Route::post('/doctores/update/{id}', [doctorController::class, 'update']);
-Route::post('/doctores/patch/{id}', [doctorController::class, 'updatePartial']);
+Route::delete('/doctores/{id}', [doctorController::class, 'destroy'])->middleware(['auth:sanctum', 'abilities:server-admin']);
+Route::post('/doctores', [doctorController::class, 'store'])->middleware(['auth:sanctum', 'abilities:server-admin']);
+Route::post('/doctores/update/{id}', [doctorController::class, 'update'])->middleware(['auth:sanctum', 'abilities:server-admin']);
+Route::post('/doctores/patch/{id}', [doctorController::class, 'updatePartial'])->middleware(['auth:sanctum', 'abilities:server-admin']);
